Refactor/remove duplication between class and function extension (#1)
* refactor(enhanceComponent): Remove duplicate code

The enhancement functions were both exporting a class-based component
with one single difference in logic. If the component was a function, it
would get executed, otherwise, it's render method would get called.

Signed-off-by: Adrian Oprea <adrian@oprea.rocks>

* fix(enhanceComponent): Inherit from the right component

In the case of functional components we should not inherit from the
component itself. We should inherit form React.Component. The validation
allows us to extend from the correct component without sacrificing
modularity

Signed-off-by: Adrian Oprea <adrian@oprea.rocks>